Create README.md
Recent graduate seeking a challenging role in IT Company. Proven leadership skills and strong academic background. Actively involved in extracurricular activities, serving as a core member of the Entrepreneurship Cell and Training & Placement Cell. Demonstrated ability to collaborate and represent peers as Class Representative for four years. Passionate about Coding and getting an IT Company Job. Eager to contribute to a dynamic team and make a meaningful impact.